Avoid split packages in Flyway extension
(cherry picked from commit e0f9964af1db4792f4861e358f4c13a67f27886f)
