+++ extensions/flyway/deployment/src/main/java/io/quarkus/flyway/deployment/FlywayEnabled.java
@@ -1,4 +1,4 @@
-package io.quarkus.flyway;
+package io.quarkus.flyway.deployment;
 
 import java.util.function.BooleanSupplier;
